django_docker/app/requirements.txt
asgiref==3.5.2
Django==3.2.6
psycopg2-binary==2.9.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
pytz==2022.1
sqlparse==0.4.2
drf-spectacular==0.21.0;
django-filter==2.4.0;

